add init_frame

+++ sperl_frame.h
@@ -4,8 +4,8 @@
 #include "sperl.h"
 
 struct SPerl_frame {
-  int32_t* operand_stack;
-  int32_t* call_stack;
+  int64_t* operand_stack;
+  int64_t* call_stack;
 };
 
 SPerl_OP* SPerl_FRAME_newOP(SPerl* sperl);
